Refactor AccountTypeSelection component and update registration form
The AccountTypeSelection component has been refactored enhancing its functionality by adding specific click handlers for business and personal accounts. This change allows for a dynamic change of the post URL in the registration form. Additionally, Inertia's useForm has been implemented in the Register component to handle form data and state, replacing the useState previously used for data. This is in line with best practices for handling form submissions in Inertia. The overall layout remains unaffected with these changes.