Add functionality to update jobs posts
The update functionality is added to the JobsController and the corresponding routes are updated as well. Changes in the frontend allow the form to display the current job details and send these details via a POST request to the update route. The visual presentation of the job location is also updated in the JobDetails component to use the formatText function.