Add coordinates field to models and relevant migrations
This commit introduces a new 'coordinates' field to the User, Company, and CompanyJob models in our Laravel application. It also includes the corresponding database migration files to add this new field to the relevant database tables. A third-party library (Matan Yadaev's Eloquent Spatial extension) is used to recognize and handle geographic coordinates.