Add admin attribute and company verification migrations
Two database migrations are added in this commit. First one introduces an 'admin' boolean attribute to the 'users' table with a default value of 'false'. The second one brings a nullable timestamp field 'company_verified_at' to the 'companies' table. In addition, an AdminPanelProvider for Filament has been created. Filament css file initialized.