Add Symfony packages and configure Postmark mailer
The Symfony `http-client` and `postmark-mailer` packages have been added to the project. Postmark mailer is now set as the default mailer and has been configured with a timeout and message stream id. Furthermore, the `POSTMARK_TOKEN` has been added to the `.env.example` file in preparation for Postmark API integration.